requirements.txt
fastapi
uvicorn
appwrite
python-dotenv
razorpay
slowapi
gunicorn
